+++ src/components/CircleColor/CircleColor.scss
