+++ src/components/index.js
@@ -1,6 +1,6 @@
 import ButtonOfPage from './Button/ButtonOfPage';
 import ButtonSize from './ButtonSize/ButtonSize';
-import CricleColor from './CricleColor/CricleColor';
+import CircleColor from './CircleColor/CircleColor';
 import Footer from './Footer/Footer';
 import AddProductForm from './Form/AddProductForm/AddProductForm';
 import Navbar from './Navbar/Navbar';
@@ -10,7 +10,7 @@ export {
   ButtonOfPage,
   AddProductForm,
   Navbar,
-  CricleColor,
+  CircleColor,
   ButtonSize,
   Footer,
   ProductDetailModal,
