+++ src/main.jsx
@@ -4,8 +4,8 @@ import { Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
 
 import App from './App';
-import { store } from './app/store';
 import './index.scss';
+import { store } from './store/store';
 import './style/customize-theme.css';
 
 ReactDOM.createRoot(document.getElementById('root')).render(
