+++ src/pages/Product/ListProduct/index.jsx
@@ -1,14 +1,14 @@
 import { Col, Row } from 'antd';
 import React, { useEffect, useState } from 'react';
 
-import instance from '../../../utils/AxiosConfig/AxiosConfig';
+import { axiosInstance } from '../../../utils/AxiosConfig/AxiosConfig';
 import SingleProduct from './SingleProduct';
 import './styles.scss';
 
 const ListProduct = () => {
   const [products, setProducts] = useState();
   const getAllProduct = () => {
-    instance
+    axiosInstance
       .get('/products')
       .then(function (response) {
         const { data } = response;
